Bump FRAMEWORK_VERSION from A to B
This works around system cache issues. It seems like if an application runs and Sparkle launches an XPC Service which is launched directly from the app bundle, then a newer bundle at the same path where Sparkle embeds XPC Services will fail to launch. There were several workarounds:

* Move the app to any other location
* Reboot the system

This can be reproduced by launching a version of an app embedding XPC Services the old way, and replacing it with a version of the app embedding XPC Services the new way.

To prevent these issues from happening, without requiring the user to move the app or reboot their Mac, we will change the FRAMEWORK version so that Sparkle executable is installed to a different path.