Add new version 3 binary delta format
More work needs to be done still:
* More reliable and performant IO writing IO writing (> 2 GiB files)
* Better error tracking
* Misc code cleanup, documentation

However all tests pass and creating/applying diffs does work, which is quite a milestone. Patch files are slightly less (0.5 - 2 MB or so) than version 2 patches likely due to metadata storage and may be more performant, but there's still more room for improvement there.

We also remove the permissions validation for needing 0755 or 0644 in the after tree which was too strict.