Don't let status window make other active windows exit in stage manager (#2153)
We set the status window's collection behavior to full screen auxiliary. This is how the collection behavior for the update alerts and permission prompts are set up, but we never applied it to the status window.

With this change, bringing the status window up front (e.g, for downloading an update) does not push the other app's windows back into the stage manager tiles.