Wait for appcast item registration before finishing update cycle
This fixes a potential race issue where sometimes the automatic update driver on completion would not trigger to prompt an update alert immediately for critical updates. Note in this case, the update would still be installed on app termination and would still be scheduled to alert the user on the regular update check interval, so this issue is not severe.