Wait for a threshold amount of time before replying with status info data (#2230)
Without this, critical update alerts may not show up as promptly as they should when automatically installing them.

This fixes a potential race issue where sometimes the automatic update driver on completion would sometimes not trigger to prompt an update alert immediately for critical updates. Note in this case, the update would still be installed on app termination and would still be scheduled to alert the user on the regular update check interval, so this issue is not severe.