Find potential matching translocated running apps in updater agent
In the installer progress agent, when we find all potential running applications for a host bundle, we should also consider any translocated running apps that look-alike the host bundle.

This matters when updating external bundles. For example if sparkle-cli tries to update an app on disk in ~/Downloads/ but a running instance of the app is translocated, sparkle-cli should terminate the translocated instance of the app before installing the update.

It's possible this could generate a false positive but we err on the side of safety here. We only use running translocated apps if we find no regular matching apps.