Perform Gatekeeper scan of new update on macOS Sonoma
Performing a Gatekeeper scan early on can remove the need for the system to perform the scan and bring up a verification dialog on the update's first launch.

Also move a lot of the installation work to be done during the first phase of installation (when possible), before the target app needs to be terminated. For the common case where the old and new bundles are on the same volume, the final phase of installation requiring a restart is just an atomic swap.