Fix Updater.app not starting when running Sparkle as root on Sonoma
This may occur for example when you run sparkle-cli as root or use Sparkle.framework in a root process. This is not typically a common use case, but could be used from daemons.

The issue is that Sonoma doesn't like starting our Updater app in the user domain when intermediate directories containing it are root owned. So now we try to create intermediate directories in the standard user's home directory.